requirements.txt
flask
docker
gitpython